Transaction signing functionality implemented. Tranaction validation functionality implemented. New hash function for mining a block was implemented replacing openssl. Blockchain class and functionality complete. Commit before submission.

+++ src/classes/transactions/SignedTransaction.cpp
@@ -41,10 +41,11 @@ string SignedTransaction::getSender() const {
 string SignedTransaction::getReceiver() const {
     return receiver;
 }
-int SignedTransaction::getAmount() const {
+double SignedTransaction::getAmount() const {
     return amount;
 }
 
 time_t SignedTransaction::getTimeStamp() const {
     return timestamp;
 }
+
